reverse-transaction-service/requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
google-generativeai
